# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CloudForge - Visual Cloud Infrastructure Design Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Project {
  id: number;
  name: string;
  description?: string | null;
  /** aws | azure | gcp | multi */
  provider: string;
  componentCount: number;
  connectionCount: number;
  createdAt: string;
  updatedAt: string;
}

/**
 * Provider-specific configuration JSON
 */
export type ComponentConfig = { [key: string]: unknown };

export interface Component {
  id: number;
  projectId: number;
  /** vpc | subnet | ec2 | rds | s3 | elb | eks | lambda | apigateway | cloudfront | route53 | sqs | sns | iam | nat_gateway | internet_gateway | security_group | virtual_network | vm | sql_database | blob_storage | aks | functions | app_gateway | key_vault | vnet | gke | cloud_sql | cloud_storage | cloud_functions | cloud_run | bigquery | pub_sub | load_balancer | custom */
  type: string;
  /** aws | azure | gcp */
  provider: string;
  name: string;
  /** Provider-specific configuration JSON */
  config: ComponentConfig;
  positionX: number;
  positionY: number;
  width?: number | null;
  height?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Connection {
  id: number;
  projectId: number;
  sourceComponentId: number;
  targetComponentId: number;
  label?: string | null;
  /** network | dependency | data_flow | security */
  connectionType?: string | null;
  sourceHandle?: string | null;
  targetHandle?: string | null;
  createdAt: string;
}

export interface ProjectDetail {
  id: number;
  name: string;
  description?: string | null;
  provider: string;
  components: Component[];
  connections: Connection[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  description?: string | null;
  provider: string;
}

export interface UpdateProjectBody {
  name?: string;
  description?: string | null;
  provider?: string;
}

export type ProjectStatsComponentsByType = { [key: string]: number };

export type ProjectStatsComponentsByProvider = { [key: string]: number };

export interface ProjectStats {
  totalComponents: number;
  totalConnections: number;
  componentsByType: ProjectStatsComponentsByType;
  componentsByProvider: ProjectStatsComponentsByProvider;
}

export type CreateComponentBodyConfig = { [key: string]: unknown };

export interface CreateComponentBody {
  type: string;
  provider: string;
  name: string;
  config?: CreateComponentBodyConfig;
  positionX: number;
  positionY: number;
  width?: number | null;
  height?: number | null;
}

export type UpdateComponentBodyConfig = { [key: string]: unknown };

export interface UpdateComponentBody {
  name?: string;
  config?: UpdateComponentBodyConfig;
  positionX?: number;
  positionY?: number;
  width?: number | null;
  height?: number | null;
}

export interface CreateConnectionBody {
  sourceComponentId: number;
  targetComponentId: number;
  label?: string | null;
  connectionType?: string | null;
  sourceHandle?: string | null;
  targetHandle?: string | null;
}

export interface ExportBody {
  /** terraform | cloudformation | kubernetes | yaml */
  format: string;
}

export interface ExportResult {
  format: string;
  content: string;
  filename: string;
}

export interface ValidationIssue {
  /** error | warning */
  severity: string;
  componentId?: number | null;
  message: string;
  code: string;
}

export interface ValidationSuggestion {
  /** high_availability | cost | security | performance */
  type: string;
  message: string;
  componentId?: number | null;
}

export interface ValidationResult {
  valid: boolean;
  issues: ValidationIssue[];
  suggestions: ValidationSuggestion[];
}

export interface Template {
  id: number;
  name: string;
  description: string;
  /** web_app | microservices | serverless | data_pipeline | kubernetes | networking */
  category: string;
  provider: string;
  componentCount: number;
  tags: string[];
}

export type TemplateDetailComponentsItem = { [key: string]: unknown };

export type TemplateDetailConnectionsItem = { [key: string]: unknown };

export interface TemplateDetail {
  id: number;
  name: string;
  description: string;
  category: string;
  provider: string;
  components: TemplateDetailComponentsItem[];
  connections: TemplateDetailConnectionsItem[];
  tags: string[];
}

export interface ImportTerraformBody {
  /** Terraform HCL file content */
  content: string;
  /** Optional project name override */
  projectName?: string | null;
}

export interface ImportTerraformResult {
  project: Project;
  componentCount: number;
  connectionCount: number;
}

export interface ChatMessage {
  /** user | assistant */
  role: string;
  content: string;
}

export interface AiChatBody {
  message: string;
  projectId?: number | null;
  history?: ChatMessage[];
}

export interface AiChatResponse {
  message: string;
  suggestions?: string[] | null;
}

export type UseTemplate201 = {
  project: ProjectSummary;
};