# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CloudForge - Visual Cloud Infrastructure Design Tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/components:
    get:
      operationId: listComponents
      tags: [components]
      summary: List all components in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Component"
    post:
      operationId: createComponent
      tags: [components]
      summary: Add a component to the canvas
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComponentBody"
      responses:
        "201":
          description: Created component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
  /projects/{projectId}/components/{componentId}:
    put:
      operationId: updateComponent
      tags: [components]
      summary: Update a component
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComponentBody"
      responses:
        "200":
          description: Updated component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
    delete:
      operationId: deleteComponent
      tags: [components]
      summary: Delete a component
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List all connections in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      operationId: createConnection
      tags: [connections]
      summary: Create a connection between two components
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionBody"
      responses:
        "201":
          description: Created connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
  /projects/{projectId}/connections/{connectionId}:
    delete:
      operationId: deleteConnection
      tags: [connections]
      summary: Delete a connection
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: connectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/export:
    post:
      operationId: exportProject
      tags: [export]
      summary: Export project as IaC (Terraform, CloudFormation, Kubernetes)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportBody"
      responses:
        "200":
          description: Generated IaC output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"
...
components:
  schemas:
...
    Component:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
          description: "vpc | subnet | ec2 | rds | s3 | elb | eks | lambda | apigateway | cloudfront | route53 | sqs | sns | iam | nat_gateway | internet_gateway | security_group | virtual_network | vm | sql_database | blob_storage | aks | functions | app_gateway | key_vault | vnet | gke | cloud_sql | cloud_storage | cloud_functions | cloud_run | bigquery | pub_sub | load_balancer | custom"
        provider:
          type: string
          description: "aws | azure | gcp"
        name:
          type: string
        config:
          type: object
          description: Provider-specific configuration JSON
          additionalProperties: true
        positionX:
          type: number
        positionY:
          type: number
        width:
          type: number
          nullable: true
        height:
          type: number
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - type
        - provider
        - name
        - config
        - positionX
        - positionY
        - createdAt
        - updatedAt
    CreateComponentBody:
      type: object
      properties:
        type:
          type: string
        provider:
          type: string
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        positionX:
          type: number
        positionY:
          type: number
        width:
          type: number
          nullable: true
        height:
          type: number
          nullable: true
      required:
        - type
        - provider
        - name
        - positionX
        - positionY
    UpdateComponentBody:
      type: object
      properties:
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        positionX:
          type: number
        positionY:
          type: number
        width:
          type: number
          nullable: true
        height:
          type: number
          nullable: true
    Connection:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        sourceComponentId:
          type: integer
        targetComponentId:
          type: integer
        label:
          type: string
          nullable: true
        connectionType:
          type: string
          description: "network | dependency | data_flow | security"
          nullable: true
        sourceHandle:
          type: string
          nullable: true
        targetHandle:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - sourceComponentId
        - targetComponentId
        - createdAt
    CreateConnectionBody:
      type: object
      properties:
        sourceComponentId:
          type: integer
        targetComponentId:
          type: integer
        label:
          type: string
          nullable: true
        connectionType:
          type: string
          nullable: true
        sourceHandle:
          type: string
          nullable: true
        targetHandle:
          type: string
          nullable: true
      required:
        - sourceComponentId
        - targetComponentId
    ExportBody:
      type: object
      properties:
        format:
          type: string
          description: "terraform | cloudformation | kubernetes | yaml"
      required:
        - format
    ExportResult:
      type: object
      properties:
        format:
          type: string
        content:
          type: string
        filename:
          type: string
      required:
        - format
        - content
        - filename
...